Move files that are shared between Lobby and Game
