+++ src/InterfaceAdapters/CommandExecutor/CommandDTO.ts
