+++ src/InterfaceAdapters/CommandExecutor/GameCommands/Interfaces/ICard.ts
