+++ src/InterfaceAdapters/CommandExecutor/GameCommands/Interfaces/IPlayer.ts
