+++ src/InterfaceAdapters/CommandExecutor/GameCommands/PlayAgainCommand.ts
@@ -1,5 +1,5 @@
-import IGame from '../Interfaces/IGame'
-import ICommand from './ICommand'
+import IGame from './Interfaces/IGame'
+import ICommand from '../ICommand'
 
 class PlayAgainCommand implements ICommand {
   constructor(private readonly game: IGame) {}
