+++ src/InterfaceAdapters/CommandExecutor/ICommand.ts
