+++ src/InterfaceAdapters/CommandExecutor/LobbyCommands/LobbyCommandDTOs/HostNewGameCommandDTO.ts
@@ -1,4 +1,4 @@
-import CommandDTO from '../../GameCommandDTOs/CommandDTO'
+import CommandDTO from '../../CommandDTO'
 
 interface HostNewGameCommandDTO extends CommandDTO {
   name: 'hostNewGame'
