+++ src/InterfaceAdapters/CommandExecutor/LobbyCommands/LobbyCommandFactory.ts
@@ -1,6 +1,6 @@
-import CommandDTO from '../GameCommandDTOs/CommandDTO'
-import ICommand from '../GameCommands/ICommand'
-import ICommandFactory from '../GameCommands/ICommandFactory'
+import CommandDTO from '../CommandDTO'
+import ICommand from '../ICommand'
+import ICommandFactory from '../ICommandFactory'
 import HostNewGameCommand from './HostNewGameCommand'
 import IGameManager from './IGameManager'
 import HostNewGameCommandDTO from './LobbyCommandDTOs/HostNewGameCommandDTO'
