+++ src/InterfaceAdapters/LocalGameCommandInterface.ts
@@ -1,9 +1,9 @@
-import CommandDTO from './CommandExecutor/GameCommandDTOs/CommandDTO'
-import GameCommandExecutor from './CommandExecutor/GameCommandExecutor'
+import CommandDTO from './CommandExecutor/CommandDTO'
+import CommandExecutor from './CommandExecutor/CommandExecutor'
 import ICommandInterface from './ICommandInterface'
 
 class LocalGameCommandInterface implements ICommandInterface {
-  constructor(private readonly gameCommandExecutor: GameCommandExecutor) {}
+  constructor(private readonly gameCommandExecutor: CommandExecutor) {}
 
   public async giveCommand(command: CommandDTO): Promise<void> {
     this.gameCommandExecutor.execute(command)
